requirements114.txt
absl_py==1.2.0
jax==0.3.24
jaxlib==0.3.24
ml_collections==0.1.0
pandas==1.4.2
Pillow==9.2.0
requests==2.27.1
six==1.15.0
tensorflow==2.9.0
tensorflow-probability==0.17.0
tensorflow_datasets==4.5.2
tensorflow_gan==2.0.0
tensorflow_hub==0.12.0
tqdm==4.64.0
ninja
